+++ backend/src/main/kotlin/ch/zhaw/pm4/simonsays/service/MenuService.kt
@@ -5,6 +5,7 @@ import ch.zhaw.pm4.simonsays.api.types.MenuCreateUpdateDTO
 import ch.zhaw.pm4.simonsays.api.types.MenuDTO
 import ch.zhaw.pm4.simonsays.entity.Menu
 import ch.zhaw.pm4.simonsays.entity.MenuItem
+import ch.zhaw.pm4.simonsays.exception.ResourceInUseException
 import ch.zhaw.pm4.simonsays.exception.ResourceNotFoundException
 import ch.zhaw.pm4.simonsays.repository.MenuItemRepository
 import ch.zhaw.pm4.simonsays.repository.MenuRepository
@@ -43,6 +44,9 @@ class MenuService(
 
     fun deleteMenu(menuId: Long, eventId: Long) {
         val menu = getMenuEntity(menuId, eventId)
+        if (!menu.orderMenu.isNullOrEmpty()) {
+            throw ResourceInUseException("Menu is used in orders and cannot be deleted")
+        }
         menuRepository.delete(menu)
     }
 
@@ -60,5 +64,4 @@ class MenuService(
         return menuRepository.findByIdAndEventId(menuId, eventId)
                 .orElseThrow { ResourceNotFoundException("Menu not found with ID: $menuId") }
     }
-
 }
